vue-pizza: refractor HomeView into constructor module
