+++ frontend/src/common/components/AppDrop.vue
@@ -6,21 +6,21 @@
 </template>
 
 <script setup>
-import {reactive} from 'vue';
+  import {reactive} from 'vue';
 
-import {events} from '../events';
-import {eventEmitter} from '../services/event-bus';
+  import {events} from '../events';
+  import {eventEmitter} from '../services/event-bus';
 
-const items = reactive([]);
+  const items = reactive([]);
 
-const onDragOver = (event) => {
-  event.dataTransfer.dropEffect = 'move';
-};
+  const onDragOver = (event) => {
+    event.dataTransfer.dropEffect = 'move';
+  };
 
-const onDrop = (event) => {
-  const data = event.dataTransfer.getData('application/json');
-  if (data) {
-    eventEmitter.emit(events.DROP, {data: JSON.parse(data)});
-  }
-};
+  const onDrop = (event) => {
+    const data = event.dataTransfer.getData('application/json');
+    if (data) {
+      eventEmitter.emit(events.DROP, {data: JSON.parse(data)});
+    }
+  };
 </script>
