+++ frontend/src/layouts/index.js
@@ -1,0 +1,3 @@
+import ContentLayout from "./ContentLayout.vue";
+
+export {ContentLayout};
