+++ frontend/src/modules/constructor/index.js
@@ -1,0 +1,7 @@
+import SectionDough from "./SectionDough.vue";
+import SectionIngredients from "./SectionIngredients.vue";
+import SectionPizza from "./SectionPizza.vue";
+import SectionSize from "./SectionSize.vue";
+import SectionSauce from "./SectionSauce.vue";
+
+export {SectionDough, SectionIngredients, SectionPizza, SectionSize, SectionSauce};
